// Front/ProcesoMatricula/src/app/model/docente.ts
export interface Docente {
    id: number;
    cedula: string;
    nombre: string;
    apellido1: string;
    apellido2: string;
    direccion: string;
    fecha_nacimiento: string;
    telefono: string;
    correo: string;
    provinciaId: number;
    FechaDeCreacion?: string;
    ActualizadoEn?: string;
  }

  /*
export interface Docente {
  id: number;
  cedula: string;
  nombre: string;
  apellido1: string;
  apellido2: string;
  direccion: string;
  fecha_nacimiento: string;
  telefono: string;
  correo: string;
  provinciaId: number;
  FechaDeCreacion?: string;
  ActualizadoEn?: string;
}
  */